requirements.txt
anyio==3.3.4
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
orjson==3.6.4
packaging==21.2
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.1
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
